STM32\RealProjects\STM32_Dune
Rename software component Com to ComManager
