+++ STM32/RealProjects/STM32_Dune/SWC/COMMANAGER/ComManager.h
@@ -1,0 +1,33 @@
+#ifndef SWC_COMMANAGER_HEADER
+#define SWC_COMMANAGER_HEADER
+
+/*--------------------------------------------------------------------------------
+*                           Libraries
+*--------------------------------------------------------------------------------*/
+#include "GeneralTypes.h" 
+
+/*--------------------------------------------------------------------------------
+*                           Custom data types
+*--------------------------------------------------------------------------------*/
+
+/*--------------------------------------------------------------------------------
+*                           Const and Macro
+*--------------------------------------------------------------------------------*/
+
+/* This is the main activation switch of "ComManager" software component */
+#define SWC_COMMANAGER_ENABLE       ON
+
+/* "ComManager" task period [ms] */
+#define SWC_COMMANAGER_PERIOD       5U
+
+/*--------------------------------------------------------------------------------
+*                           Global variables
+*--------------------------------------------------------------------------------*/
+
+/*--------------------------------------------------------------------------------
+*                           Functions prototypes
+*--------------------------------------------------------------------------------*/
+void ComManager_InitRunnable(void);
+void ComManager_MainRunnable(void);
+
+#endif /* SWC_COMMANAGER_HEADER */
